+++ src/test/resources/inspections/ForeachToCollectionInspection/foreach-outer-scope-this-variable.php
@@ -1,0 +1,14 @@
+<?php
+
+$array = [
+    'one',
+    'two',
+    'tree'
+];
+
+
+$this->name = 'one';
+
+<caret>foreach ($array as $item) {
+    $this->name = $item;
+}